requirements.txt
mss
requests
Pillow
markdown
jinja2